Added new topic_fetchers file.
