+++ core/controllers/topic_editor.py
@@ -28,6 +28,7 @@ from core.domain import story_services
 from core.domain import subtopic_page_domain
 from core.domain import subtopic_page_services
 from core.domain import topic_domain
+from core.domain import topic_fetchers
 from core.domain import topic_services
 from core.domain import user_services
 import feconf
@@ -45,7 +46,7 @@ class TopicEditorStoryHandler(base.BaseHandler):
     @acl_decorators.can_view_any_topic_editor
     def get(self, topic_id):
         """Handles GET requests."""
-        topic = topic_services.get_topic_by_id(topic_id)
+        topic = topic_fetchers.get_topic_by_id(topic_id)
         canonical_story_summaries = story_services.get_story_summaries_by_ids(
             topic.canonical_story_ids)
         additional_story_summaries = story_services.get_story_summaries_by_ids(
@@ -90,7 +91,7 @@ class TopicEditorPage(base.BaseHandler):
     @acl_decorators.can_view_any_topic_editor
     def get(self, topic_id):
         """Handles GET requests."""
-        topic = topic_services.get_topic_by_id(topic_id, strict=False)
+        topic = topic_fetchers.get_topic_by_id(topic_id, strict=False)
 
         if topic is None:
             raise self.PageNotFoundException(
@@ -164,7 +165,7 @@ class EditableTopicDataHandler(base.BaseHandler):
     @acl_decorators.can_view_any_topic_editor
     def get(self, topic_id):
         """Populates the data on the individual topic page."""
-        topic = topic_services.get_topic_by_id(topic_id, strict=False)
+        topic = topic_fetchers.get_topic_by_id(topic_id, strict=False)
 
         if topic is None:
             raise self.PageNotFoundException(
@@ -192,7 +193,7 @@ class EditableTopicDataHandler(base.BaseHandler):
         includes editing its html data as of now).
         """
         topic_domain.Topic.require_valid_topic_id(topic_id)
-        topic = topic_services.get_topic_by_id(topic_id, strict=False)
+        topic = topic_fetchers.get_topic_by_id(topic_id, strict=False)
 
         version = self.payload.get('version')
         self._require_valid_version(version, topic.version)
@@ -216,7 +217,7 @@ class EditableTopicDataHandler(base.BaseHandler):
         except utils.ValidationError as e:
             raise self.InvalidInputException(e)
 
-        topic = topic_services.get_topic_by_id(topic_id, strict=False)
+        topic = topic_fetchers.get_topic_by_id(topic_id, strict=False)
         skill_ids = topic.get_all_skill_ids()
 
         skill_id_to_description_dict = (
@@ -233,7 +234,7 @@ class EditableTopicDataHandler(base.BaseHandler):
     def delete(self, topic_id):
         """Handles Delete requests."""
         topic_domain.Topic.require_valid_topic_id(topic_id)
-        topic = topic_services.get_topic_by_id(topic_id, strict=False)
+        topic = topic_fetchers.get_topic_by_id(topic_id, strict=False)
         if topic is None:
             raise self.PageNotFoundException(
                 'The topic with the given id doesn\'t exist.')
@@ -296,7 +297,7 @@ class TopicPublishHandler(base.BaseHandler):
     @acl_decorators.can_change_topic_publication_status
     def put(self, topic_id):
         """Publishes or unpublishes a topic."""
-        topic = topic_services.get_topic_by_id(topic_id, strict=False)
+        topic = topic_fetchers.get_topic_by_id(topic_id, strict=False)
         if topic is None:
             raise self.PageNotFoundException
