+++ core/controllers/topic_editor_test.py
@@ -17,6 +17,7 @@
 from core.domain import skill_services
 from core.domain import story_services
 from core.domain import topic_domain
+from core.domain import topic_fetchers
 from core.domain import topic_services
 from core.domain import user_services
 from core.tests import test_utils
@@ -50,7 +51,7 @@ class BaseTopicEditorControllerTests(test_utils.GenericTestBase):
         self.skill_id_2 = skill_services.get_new_skill_id()
         self.save_new_skill(
             self.skill_id_2, self.admin_id, 'Skill Description 2')
-        self.topic_id = topic_services.get_new_topic_id()
+        self.topic_id = topic_fetchers.get_new_topic_id()
         self.save_new_topic(
             self.topic_id, self.admin_id, 'Name', 'Description', [], [],
             [self.skill_id, self.skill_id_2], [], 1)
@@ -68,7 +69,7 @@ class TopicEditorStoryHandlerTests(BaseTopicEditorControllerTests):
     def test_handler_updates_story_summary_dicts(self):
         self.login(self.ADMIN_EMAIL)
 
-        topic_id = topic_services.get_new_topic_id()
+        topic_id = topic_fetchers.get_new_topic_id()
         canonical_story_id = story_services.get_new_story_id()
         additional_story_id = story_services.get_new_story_id()
 
@@ -136,7 +137,7 @@ class SubtopicPageEditorTests(BaseTopicEditorControllerTests):
         self.get_json(
             '%s/%s/%s' % (
                 feconf.SUBTOPIC_PAGE_EDITOR_DATA_URL_PREFIX,
-                self.topic_id, topic_services.get_new_topic_id()),
+                self.topic_id, topic_fetchers.get_new_topic_id()),
             expected_status_int=404)
 
         self.logout()
@@ -237,7 +238,7 @@ class TopicEditorTests(BaseTopicEditorControllerTests):
         self.get_html_response(
             '%s/%s' % (
                 feconf.TOPIC_EDITOR_URL_PREFIX,
-                topic_services.get_new_topic_id()), expected_status_int=404)
+                topic_fetchers.get_new_topic_id()), expected_status_int=404)
 
         self.logout()
 
@@ -294,7 +295,7 @@ class TopicEditorTests(BaseTopicEditorControllerTests):
         self.get_json(
             '%s/%s' % (
                 feconf.TOPIC_EDITOR_DATA_URL_PREFIX,
-                topic_services.get_new_topic_id()), expected_status_int=404)
+                topic_fetchers.get_new_topic_id()), expected_status_int=404)
 
         self.logout()
 
@@ -474,7 +475,7 @@ class TopicEditorTests(BaseTopicEditorControllerTests):
         # topic version.
         self.login(self.ADMIN_EMAIL)
 
-        topic_id_1 = topic_services.get_new_topic_id()
+        topic_id_1 = topic_fetchers.get_new_topic_id()
         self.save_new_topic(
             topic_id_1, self.admin_id, 'Name 1', 'Description 1', [], [],
             [self.skill_id], [], 1)
@@ -601,7 +602,7 @@ class TopicEditorTests(BaseTopicEditorControllerTests):
         self.delete_json(
             '%s/%s' % (
                 feconf.TOPIC_EDITOR_DATA_URL_PREFIX,
-                topic_services.get_new_topic_id()), expected_status_int=404)
+                topic_fetchers.get_new_topic_id()), expected_status_int=404)
 
         self.logout()
 
@@ -656,7 +657,7 @@ class TopicRightsHandlerTests(BaseTopicEditorControllerTests):
         json_response = self.get_json(
             '%s/%s' % (
                 feconf.TOPIC_RIGHTS_URL_PREFIX,
-                topic_services.get_new_topic_id()), expected_status_int=400)
+                topic_fetchers.get_new_topic_id()), expected_status_int=400)
         self.assertEqual(
             json_response['error'],
             'Expected a valid topic id to be provided.')
@@ -716,7 +717,7 @@ class TopicPublishHandlerTests(BaseTopicEditorControllerTests):
 
         csrf_token = self.get_new_csrf_token()
 
-        new_topic_id = topic_services.get_new_topic_id()
+        new_topic_id = topic_fetchers.get_new_topic_id()
         self.put_json(
             '%s/%s' % (
                 feconf.TOPIC_STATUS_URL_PREFIX, new_topic_id),
