+++ core/controllers/topic_viewer.py
@@ -18,7 +18,7 @@ from constants import constants
 from core.controllers import acl_decorators
 from core.controllers import base
 from core.domain import story_services
-from core.domain import topic_services
+from core.domain import topic_fetchers
 import feconf
 
 
@@ -48,7 +48,7 @@ class TopicPageDataHandler(base.BaseHandler):
         if not constants.ENABLE_NEW_STRUCTURE_PLAYERS:
             raise self.PageNotFoundException
 
-        topic = topic_services.get_topic_by_name(topic_name)
+        topic = topic_fetchers.get_topic_by_name(topic_name)
 
         canonical_story_summaries = [
             story_services.get_story_summary_by_id(
