+++ core/domain/topic_services.py
@@ -25,6 +25,7 @@ from core.domain import state_domain
 from core.domain import subtopic_page_domain
 from core.domain import subtopic_page_services
 from core.domain import topic_domain
+from core.domain import topic_fetchers
 from core.domain import user_services
 from core.platform import models
 import feconf
@@ -32,88 +33,6 @@ import feconf
 (topic_models,) = models.Registry.import_models([models.NAMES.topic])
 datastore_services = models.Registry.import_datastore_services()
 memcache_services = models.Registry.import_memcache_services()
-
-
-def _migrate_subtopics_to_latest_schema(versioned_subtopics):
-    """Holds the responsibility of performing a step-by-step, sequential update
-    of the subtopics structure based on the schema version of the input
-    subtopics dictionary. If the current subtopics schema changes, a
-    new conversion function must be added and some code appended to this
-    function to account for that new version.
-
-    Args:
-        versioned_subtopics: A dict with two keys:
-          - schema_version: int. The schema version for the subtopics dict.
-          - subtopics: list(dict). The list of dicts comprising the topic's
-              subtopics.
-
-    Raises:
-        Exception: The schema version of subtopics is outside of what
-            is supported at present.
-    """
-    subtopic_schema_version = versioned_subtopics['schema_version']
-    if not (1 <= subtopic_schema_version
-            <= feconf.CURRENT_SUBTOPIC_SCHEMA_VERSION):
-        raise Exception(
-            'Sorry, we can only process v1-v%d subtopic schemas at '
-            'present.' % feconf.CURRENT_SUBTOPIC_SCHEMA_VERSION)
-
-    while (subtopic_schema_version <
-           feconf.CURRENT_SUBTOPIC_SCHEMA_VERSION):
-        topic_domain.Topic.update_subtopics_from_model(
-            versioned_subtopics, subtopic_schema_version)
-        subtopic_schema_version += 1
-
-
-# Repository GET methods.
-def _get_topic_memcache_key(topic_id, version=None):
-    """Returns a memcache key for the topic.
-
-    Args:
-        topic_id: str. ID of the topic.
-        version: int. The version of the topic.
-
-    Returns:
-        str. The memcache key of the topic.
-    """
-    if version:
-        return 'topic-version:%s:%s' % (topic_id, version)
-    else:
-        return 'topic:%s' % topic_id
-
-
-def get_topic_from_model(topic_model):
-    """Returns a topic domain object given a topic model loaded
-    from the datastore.
-
-    Args:
-        topic_model: TopicModel. The topic model loaded from the
-            datastore.
-
-    Returns:
-        topic. A Topic domain object corresponding to the given
-            topic model.
-    """
-    versioned_subtopics = {
-        'schema_version': topic_model.subtopic_schema_version,
-        'subtopics': copy.deepcopy(topic_model.subtopics)
-    }
-    if (topic_model.subtopic_schema_version !=
-            feconf.CURRENT_SUBTOPIC_SCHEMA_VERSION):
-        _migrate_subtopics_to_latest_schema(versioned_subtopics)
-    return topic_domain.Topic(
-        topic_model.id, topic_model.name,
-        topic_model.description, topic_model.canonical_story_ids,
-        topic_model.additional_story_ids, topic_model.uncategorized_skill_ids,
-        [
-            topic_domain.Subtopic.from_dict(subtopic)
-            for subtopic in versioned_subtopics['subtopics']
-        ],
-        versioned_subtopics['schema_version'],
-        topic_model.next_subtopic_id,
-        topic_model.language_code,
-        topic_model.version, topic_model.created_on,
-        topic_model.last_updated)
 
 
 def get_all_topic_summaries():
@@ -138,7 +57,9 @@ def get_all_skill_ids_assigned_to_some_topic():
     """
     skill_ids = set([])
     all_topic_models = topic_models.TopicModel.get_all()
-    all_topics = [get_topic_from_model(topic) for topic in all_topic_models]
+    all_topics = [
+        topic_fetchers.get_topic_from_model(topic)
+        for topic in all_topic_models]
     for topic in all_topics:
         skill_ids.update(topic.get_all_skill_ids())
     return skill_ids
@@ -169,72 +90,6 @@ def get_topic_summary_from_model(topic_summary_model):
     )
 
 
-def get_topic_by_id(topic_id, strict=True, version=None):
-    """Returns a domain object representing a topic.
-
-    Args:
-        topic_id: str. ID of the topic.
-        strict: bool. Whether to fail noisily if no topic with the given
-            id exists in the datastore.
-        version: int or None. The version number of the topic to be
-            retrieved. If it is None, the latest version will be retrieved.
-
-    Returns:
-        Topic or None. The domain object representing a topic with the
-        given id, or None if it does not exist.
-    """
-    topic_memcache_key = _get_topic_memcache_key(topic_id, version=version)
-    memcached_topic = memcache_services.get_multi(
-        [topic_memcache_key]).get(topic_memcache_key)
-
-    if memcached_topic is not None:
-        return memcached_topic
-    else:
-        topic_model = topic_models.TopicModel.get(
-            topic_id, strict=strict, version=version)
-        if topic_model:
-            topic = get_topic_from_model(topic_model)
-            memcache_services.set_multi({topic_memcache_key: topic})
-            return topic
-        else:
-            return None
-
-
-def get_topics_by_ids(topic_ids):
-    """Returns a list of topics matching the IDs provided.
-
-    Args:
-        topic_ids: list(str). List of IDs to get topics for.
-
-    Returns:
-        list(Topic|None). The list of topics corresponding to given ids
-            (with None in place of topic ids corresponding to deleted topics).
-    """
-    all_topic_models = topic_models.TopicModel.get_multi(topic_ids)
-    topics = [
-        get_topic_from_model(topic_model) if topic_model is not None else None
-        for topic_model in all_topic_models]
-    return topics
-
-
-def get_topic_by_name(topic_name):
-    """Returns a domain object representing a topic.
-
-    Args:
-        topic_name: str. The name of the topic.
-
-    Returns:
-        Topic or None. The domain object representing a topic with the
-        given id, or None if it does not exist.
-    """
-    topic_model = topic_models.TopicModel.get_by_name(topic_name)
-    if topic_model is None:
-        return None
-
-    topic = get_topic_from_model(topic_model)
-    return topic
-
-
 def get_topic_summary_by_id(topic_id, strict=True):
     """Returns a domain object representing a topic summary.
 
@@ -254,15 +109,6 @@ def get_topic_summary_by_id(topic_id, strict=True):
         return topic_summary
     else:
         return None
-
-
-def get_new_topic_id():
-    """Returns a new topic id.
-
-    Returns:
-        str. A new topic id.
-    """
-    return topic_models.TopicModel.get_new_id('')
 
 
 def _create_topic(committer_id, topic, commit_message, commit_cmds):
@@ -307,7 +153,7 @@ def save_new_topic(committer_id, topic):
     Raises:
         Exception. Topic with same name already exists.
     """
-    existing_topic = get_topic_by_name(topic.name)
+    existing_topic = topic_fetchers.get_topic_by_name(topic.name)
     if existing_topic is not None:
         raise Exception('Topic with name \'%s\' already exists' % topic.name)
 
@@ -341,7 +187,7 @@ def apply_change_list(topic_id, change_list):
             a list of ids of the newly created subtopics and a list of changes
             applied to modified subtopic pages.
     """
-    topic = get_topic_by_id(topic_id)
+    topic = topic_fetchers.get_topic_by_id(topic_id)
     newly_created_subtopic_ids = []
     existing_subtopic_page_ids_to_be_modified = []
     deleted_subtopic_ids = []
@@ -520,7 +366,7 @@ def _save_topic(committer_id, topic, commit_message, change_list):
     topic_model.language_code = topic.language_code
     change_dicts = [change.to_dict() for change in change_list]
     topic_model.commit(committer_id, commit_message, change_dicts)
-    memcache_services.delete(_get_topic_memcache_key(topic.id))
+    memcache_services.delete(topic_fetchers.get_topic_memcache_key(topic.id))
     topic.version += 1
 
 
@@ -621,7 +467,7 @@ def delete_story(user_id, topic_id, story_id):
         topic_id: str. The id of the topic from which to remove the story.
         story_id: str. The story to remove from the topic.
     """
-    topic = get_topic_by_id(topic_id)
+    topic = topic_fetchers.get_topic_by_id(topic_id)
     old_canonical_story_ids = copy.deepcopy(topic.canonical_story_ids)
     topic.delete_story(story_id)
     change_list = [topic_domain.TopicChange({
@@ -643,7 +489,7 @@ def add_canonical_story(user_id, topic_id, story_id):
         topic_id: str. The id of the topic to which the story is to be added.
         story_id: str. The story to add to the topic.
     """
-    topic = get_topic_by_id(topic_id)
+    topic = topic_fetchers.get_topic_by_id(topic_id)
     old_canonical_story_ids = copy.deepcopy(topic.canonical_story_ids)
     topic.add_canonical_story(story_id)
     change_list = [topic_domain.TopicChange({
@@ -690,7 +536,7 @@ def delete_topic(committer_id, topic_id, force_deletion=False):
 
     # This must come after the topic is retrieved. Otherwise the memcache
     # key will be reinstated.
-    topic_memcache_key = _get_topic_memcache_key(topic_id)
+    topic_memcache_key = topic_fetchers.get_topic_memcache_key(topic_id)
     memcache_services.delete(topic_memcache_key)
 
 
@@ -711,7 +557,7 @@ def create_topic_summary(topic_id):
     Args:
         topic_id: str. ID of the topic.
     """
-    topic = get_topic_by_id(topic_id)
+    topic = topic_fetchers.get_topic_by_id(topic_id)
     topic_summary = compute_summary_of_topic(topic)
     save_topic_summary(topic_summary)
